+++ source/include/Ostap/Peaks.h
@@ -1954,8 +1954,97 @@ namespace Ostap
       // ======================================================================
     } ;
     // ========================================================================
+    /** @class Losev 
+     *  ``Losev distribution'' - asymmetric variant of Hyperbolic secant/Sech-function
+     *  \f[ f(x;\mu,\alpha,\beta) \equiv 
+     *   \frac{A}{\mathrm{e}^{-\left|\alpha\right| (x-\mu)} + 
+     *                         \mathrm{e}^{\left|\beta\right|(x-mu)}}, \f]
+     *  where \f$ A = \frac{\left|\alpha\right|+\left|\beta\right|}{\pi}.
+     *  \sin \frac{\pi\left| \beta\right| }{\left|\alpha\right|+\left|\beta\right|}\f$. 
+     *  - Leptokurtic distribution with exponential tails 
+     *  @see Losev, A., "A new lineshape for fitting x‐ray photoelectron peaks", 
+     *           Surf. Interface Anal., 14: 845-849. doi:10.1002/sia.740141207
+     *  @see  https://doi.org/10.1002/sia.740141207
+     *  @see  https://en.wikipedia.org/wiki/Hyperbolic_secant_distribution
+     */
+    class Losev 
+    {
+    public:
+      // ======================================================================
+      /** constructor from positive parameters alpha and beta 
+       *  @param mean  \f$\mu\f$-parameter 
+       *  @param alpha \f$\alpha\f$-parameter 
+       *  @param beta  \f$\beta\f$-parameter 
+       */ 
+      Losev ( const double mu    = 0 , 
+              const double alpha = 1 , 
+              const double beta  = 1 ) ;        
+      // ======================================================================
+    public:
+      // ======================================================================
+      /// evaluate the function 
+      double operator() ( const double x ) const { return pdf ( x ) ; }
+      /// evaluate the function 
+      double pdf        ( const double x ) const ;
+      // ======================================================================
+    public:
+      // ======================================================================
+      /// parameter mu 
+      double mu    () const { return m_mu    ; }
+      /// parameter alpha
+      double alpha () const { return m_alpha ; }
+      /// parameter beta 
+      double beta  () const { return m_beta  ; }      
+      // ======================================================================
+    public:
+      // ======================================================================
+      bool setMu    ( const double mu ) ;
+      bool setAlpha ( const double mu ) ;
+      bool setBeta  ( const double mu ) ;      
+      // ======================================================================
+    public:
+      // ======================================================================
+      /// the mode of the distribution 
+      double mode () const ;
+      // ======================================================================
+    public:
+      // ======================================================================
+      /// get the tag 
+      std::size_t tag () const ;
+      // ======================================================================      
+    public :
+      // ======================================================================
+      /// get the integral \f$ \int_{-\infty}^{+\infty} f(x) dx \f$
+      double integral () const { return 1 ; }
+      /** get the integral between low and high values 
+       *  \f$ \int_{low}^{high}f(x) dx\f$
+       */
+      double integral ( const double low  , 
+                        const double high ) const ;
+      // ======================================================================      
+    private :
+      // ======================================================================
+      /// parameteter "mu"
+      double m_mu     { 0 }  ; // parameter
+      /// left exponent 
+      double  m_alpha { 1 } ; // left exponent 
+      /// right exponent 
+      double  m_beta  { 1 } ; // right exponent 
+      // ======================================================================
+    private:
+      // ======================================================================
+      /// normalization 
+      mutable double m_norm { -1 } ; // normalization 
+      // ======================================================================
+    private:
+      // ======================================================================
+      /// integration workspace
+      Ostap::Math::WorkSpace m_workspace {} ; // integration workspace
+      // ======================================================================
+    } ;  
+    // ========================================================================
     /** @class  Slash 
-     *  ``Slash''-distribution -  symmetric peak with veyr heavy tail
+     *  ``Slash''-distribution -  symmetric peak with very heavy tail
      *  @see https://en.wikipedia.org/wiki/Slash_distribution
      *  Tails arew so heavy that moments (e.g. variance) do not exist 
      */